Split test.rs into separate files.
