+++ tests/common/verifying_key_bytes.rs
@@ -1,0 +1,1 @@
+pub(crate) type VerifyingKeyBytes = [u8; 32];
